fix a bug

+++ src/qtUpdateNotifier.h
@@ -114,8 +114,8 @@ private:
 	void showToolTip( const QString&,const QString&,int ) ;
 	void showToolTip( const QString&,const QString& ) ;
 	QString nextUpdateTime( void ) ;
-	QString nextUpdateTime( int ) ;
-	QString logMsg( int ) ;
+	QString nextUpdateTime( u_int64_t ) ;
+	QString logMsg( u_int64_t ) ;
 	QString logMsg( void ) ;
 	bool m_canCloseApplication ;
 	bool m_threadIsRunning ;
